refactor(test): Split combat.test.ts into focused test modules
The original combat.test.ts file was 2604 lines. Split into:
- combatBasics.test.ts (enter, phases, end)
- combatBlocking.test.ts (blocking, swift, elemental efficiency)
- combatAttacking.test.ts (attacking, resistances, elemental enemies)
- combatDamage.test.ts (damage assignment, knockout, poison)
- combatFortification.test.ts (fortification mechanics)
- combatAbilities.test.ts (paralyze, skip attack)
- combatModifiers.test.ts (armor modifiers, fame tracking)
- combatIntegration.test.ts (full combat flow)
- combatTestHelpers.ts (shared utilities)

Also fix pre-existing lint errors:
- Remove unused imports in shared/enemies/violet.ts
- Remove unused hasPendingBlock in declareBlockCommand.ts
- Remove unused imports in client DnDContext.tsx
- Remove unused isOverAnyEnemy in client EnemyCard.tsx
